move middleware in src
